# Improve performance of icon rendering in completer