--- a/packages/completion-theme/package.json
+++ b/packages/completion-theme/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@krassowski/completion-theme",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "description": "Completion theme manager for JupyterLab-LSP",
   "keywords": [
     "jupyter",