## References

Relates to #272.

## Code changes

Bindprops early to avoid cache invalidation of LabIcon. JupyterLab since 2.1 has an ability to cache LabIcon svg strings (https://github.com/jupyterlab/jupyterlab/pull/8125) but we were not getting these benefits as using bindprops late in the code was causing a react proxy to be created and used which is preventing the caching mechanism from working.
 
## User-facing changes

The rendering time for long lists of completion items decrease notably:
- `pandas.<tab>` from 178 ms to 97ms (for list of 150 completions)
- `numpy.<tab>` from 1043 ms to 641 ms (for list of 1125 completions)

We should be able to improve this further with JupyterLab 3.1 since some of the current workarounds won't be needed any longer.

#### Pandas

Before:

![Screenshot from 2021-06-15 23-35-12](https://user-images.githubusercontent.com/5832902/122138834-75d3ac00-ce3f-11eb-96c1-b0776729b3c2.png)

After:

![Screenshot from 2021-06-15 23-33-54](https://user-images.githubusercontent.com/5832902/122138842-78ce9c80-ce3f-11eb-8a62-1dabcb57f5e6.png)

### Numpy

Before

![Screenshot from 2021-06-16 01-13-03](https://user-images.githubusercontent.com/5832902/122139308-994b2680-ce40-11eb-9ff7-5373c6a66e38.png)

After

![Screenshot from 2021-06-16 01-18-17](https://user-images.githubusercontent.com/5832902/122139427-e62efd00-ce40-11eb-843b-df2832220311.png)


## Backwards-incompatible changes

None

## Chores

- [x] linted <!-- Required: Run "jlpm lint" and "python scripts/lint.py" from the root of the repository, then check this box like this: [x] -->
- [ ] tested <!-- Recommended: Let us know if you already added a test case (if relevant). -->
- [ ] documented <!-- Optional: Would it be good to improve the documentation? If yes, please consider doing this and checking this box. -->
- [x] changelog entry <!-- Recommended: Add a note in the CHANGELOG.md file under the most recent >unreleased< version; if one does not exist, feel free to create one by increasing the version number (no worries if you are not certain of the details - we can improve it later; let's just have something to work with) -->